Walking back "Pets" section
This followed some fairly significant changes that were proposed by multiple parties. I'll be creating, for posterity, a separate branch with the changes that we ultimately decided not to adopt.

For now, the end result of all the discussion was that we recognize no norm exists. Rather than remove the section entirely however, and risk people wondering what happened, the norms group decided to affirmatively state the situation. This does include the change to the name of the section, however.