Updates norm about scheduling meeting spaces
After a question about the recommended practice around impromptu use of meeting spaces, the group decided to clarify that _all_ use should be indicated on Outlook. This will prevent conflicts where multiple groups see an apparently unused space on the calendar and attempt to use it informally - particularly for people coming from outside Cubespace.

Eventually this impromptu use will be better facilitated by panels at the doors to enable immediate scheduling - but those have not yet been set up. Documentation of this will be added to this document when it occurs.